修复权限列表不显示数据

+++ user/src/main/java/com/lingyun/user/service/MenuService.java
@@ -1,13 +1,14 @@
 package com.lingyun.user.service;
 
 import com.lingyun.user.entity.MenuEntity;
+import com.lingyun.user.vo.MenuVo;
 
 import java.util.List;
 
 public interface MenuService  extends BaseService<MenuEntity>{
     public List<MenuEntity> findAll();
 
-    public List<MenuEntity> findAllByRoleId(Long roleId);
+    public List<MenuEntity> findAllByRoleId(Integer roleId);
 
 
 }
